+++ src/test/suite/index.ts
@@ -1,5 +1,5 @@
-import * as path from 'path';
-import * as Mocha from 'mocha';
+import { resolve } from 'path';
+import * as mocha from 'mocha';
 import * as glob from 'glob';
 
 export function run(): Promise<void> {
@@ -9,7 +9,7 @@ export function run(): Promise<void> {
 		color: true
 	});
 
-	const testsRoot = path.resolve(__dirname, '..');
+	const testsRoot = resolve(__dirname, '..');
 
 	return new Promise((c, e) => {
 		glob('**/**.test.js', { cwd: testsRoot }, (err, files) => {
@@ -18,7 +18,7 @@ export function run(): Promise<void> {
 			}
 
 			// Add files to the test suite
-			files.forEach(f => mocha.addFile(path.resolve(testsRoot, f)));
+			files.forEach(f => mocha.addFile(resolve(testsRoot, f)));
 
 			try {
 				// Run the mocha test